requirements.txt
ucloud-sdk-python3
